+++ src/main/java/net/geforcemods/securitycraft/items/UniversalKeyChangerItem.java
@@ -36,12 +36,12 @@ public class UniversalKeyChangerItem extends Item {
 	public InteractionResult onItemUseFirst(ItemStack stack, UseOnContext ctx) {
 		Player player = ctx.getPlayer();
 		InteractionHand hand = ctx.getHand();
-		InteractionResult briefcaseResult = handleBriefcase(player, hand).getResult();
+		Level level = ctx.getLevel();
+		InteractionResult briefcaseResult = handleBriefcase(level, player, hand).getResult();
 
 		if (briefcaseResult != InteractionResult.PASS)
 			return briefcaseResult;
 
-		Level level = ctx.getLevel();
 		BlockPos pos = ctx.getClickedPos();
 		BlockEntity be = level.getBlockEntity(pos);
 
@@ -65,10 +65,10 @@ public class UniversalKeyChangerItem extends Item {
 
 	@Override
 	public InteractionResultHolder<ItemStack> use(Level level, Player player, InteractionHand hand) {
-		return handleBriefcase(player, hand);
+		return handleBriefcase(level, player, hand);
 	}
 
-	private InteractionResultHolder<ItemStack> handleBriefcase(Player player, InteractionHand hand) {
+	private InteractionResultHolder<ItemStack> handleBriefcase(Level level, Player player, InteractionHand hand) {
 		ItemStack keyChanger = player.getItemInHand(hand);
 
 		if (hand == InteractionHand.MAIN_HAND && player.getOffhandItem().getItem() == SCContent.BRIEFCASE.get()) {
@@ -78,7 +78,7 @@ public class UniversalKeyChangerItem extends Item {
 				CompoundTag tag = briefcase.getTag();
 
 				if (tag != null && tag.contains("passcode")) {
-					if (tag.contains("saltKey"))
+					if (tag.contains("saltKey") && !level.isClientSide)
 						SaltData.removeSalt(tag.getUUID("saltKey"));
 
 					PasscodeUtils.filterPasscodeAndSaltFromTag(tag);